Better static analysis

+++ skeleton/BOOT/miyoo/app/my282.sh
@@ -10,7 +10,7 @@ MIYOO_DIR=miyoo
 
 # .tmp_update/updater does the actual installation (and later, updating)
 cp -rf .tmp_update $SDCARD_PATH/
-rm -rf "$SDCARD_PATH/$MIYOO_DIR"
+rm -rf "${SDCARD_PATH:?}/${MIYOO_DIR:?}"
 sync
 $SDCARD_PATH/.tmp_update/updater
